requirements.txt
pandas
numpy
discord.py
flask
requests
python-dotenv
mojito2
yfinance
pandas_ta
finance-datareader
tradingview_ta
investpy
bs4
datetime
pyyaml
typing
matplotlib
asyncio
glob
plotly
tabulate
yahoo_earnings_calendar
yahoo_fin
requests_html
mplfinance